backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.3
django-environ>=0.11
python-dotenv>=1.0
mysqlclient>=2.2
mysql-connector-python>=8.0
chromadb>=0.5
sentence-transformers>=2.2
python-json-logger>=2.0
PyPDF2>=3.0
python-docx>=1.1
tqdm>=4.66

# === MODELOS LLM POTENTES PARA ANÁLISIS AVANZADO ===
# Groq API (Llama 3.3 70B - PRINCIPAL, YA TIENES API KEY)
groq>=0.31.0

# Together AI (DeepSeek V3, Qwen2.5 - OPCIONAL)
together>=1.3.0

# Google Gemini (Análisis superior - OPCIONAL)  
google-generativeai>=0.8.0

# OpenAI (Backup confiable - OPCIONAL)
openai>=1.54.0
